Small package refactor.
